// apps/server/src/modules/accountUser/types/createAccountUser.accountUser.types.ts
export type CreateAccountUserInput = {
  name: string;
  cpfCnpj: string;
  password: string;
  email: string;
  phone: string;
  mobilePhone: string;
  address: string;
  addressNumber: string;
  complement: string;
  province: string;
  postalCode: string;
  externalReference: string;
  notificationDisabled: boolean;
  additionalEmails: string[];
  municipalInscription: string;
  stateInscription: string;
  observations: string;
  groupName: string;
  company: string;
};